+++ src/main.rs
@@ -2,12 +2,13 @@ use std::time::Instant;
 use chrono::{DateTime, NaiveDateTime};
 use log::info;
 use crate::traits::AsDataFrame;
-use crate::backtesting::BacktestingRunner;
+use crate::backtesting::{BacktestingConfig, BacktestingRunner};
 use crate::portfolio::PortfolioArgs;
 
 use polars::prelude::*;
 use rust_decimal_macros::dec;
 use crate::markets::utils::save_candles;
+use crate::utils::print_candle_statistics;
 
 mod backtesting;
 mod indicators;
@@ -38,25 +39,9 @@ fn main() {
         strategies::Consensus::Unison,
     );
 
-    let portfolio_args = PortfolioArgs {
-        assets: dec!(0.0),
-        capital: dec!(100.0),
-        threshold: dec!(0.0),
-        ..Default::default()
-    };
-    let manager_config = manager::PositionManagerConfig {
-        max_position_size: dec!(100.0),
-        stop_loss_percentage: dec!(0.05),
-        take_profit_percentage: dec!(0.1),
-        max_beta: dec!(1.4),
-        var_limit: dec!(10.0),
-        min_sharpe_ratio: dec!(0.4),
-        ..Default::default()
-    };
-    let mut runner = BacktestingRunner::new(
-        strategy,
-        portfolio_args,
-        manager_config,
+    let mut runner = BacktestingRunner::from_config(
+        "data/backtesting_config.toml",
+        strategy
     );
 
     let start_time = Instant::now();
@@ -64,25 +49,7 @@ fn main() {
     let performance = runner.run(&candles, &market_data).unwrap();
     let elapsed = start_time.elapsed();
 
-    // print candle statistics
-    let candle_start = candles.column("time")
-        .unwrap()
-        .datetime()
-        .unwrap()
-        .head(Some(1))
-        .get(0)
-        .unwrap();
-    let candle_start = DateTime::from_timestamp_millis(candle_start).unwrap().naive_utc();
-    let candle_end = candles.column("time")
-        .unwrap()
-        .datetime()
-        .unwrap()
-        .tail(Some(1))
-        .get(0)
-        .unwrap();
-    let candle_end = DateTime::from_timestamp_millis(candle_end).unwrap().naive_utc();
-
-    info!("Candles range: {:?} - {:?}", candle_start, candle_end);
+    print_candle_statistics(&candles);
 
     let candle_len = candles.height();
     println!("Finished processing {:?} rows in {:?}", candle_len, elapsed);
@@ -95,9 +62,10 @@ fn main() {
 
     info!("Saving data to CSV");
 
+    let save_path = "data";
+
     let candle_path = format!("data/{}.csv", candle_table);
     let market_data_path = format!("data/{}.csv", market_data_table);
-    let bbands_path = "data/bbands_graph.csv";
 
     // create mutable versions of the DataFrames
     let mut candles = candles.clone();
@@ -108,6 +76,5 @@ fn main() {
     save_candles(&mut market_data, &market_data_path).unwrap();
 
     // save indicator graph as CSV
-    let mut bbands = runner.get_strategy_as_mut().indicators[0].as_mut();
-    let bbands_graph = bbands.save_graph_as_csv(bbands_path).unwrap();
+    runner.save_indicator_data(save_path).unwrap()
 }